+++ circuits/examples/json_verification.rs
@@ -4,7 +4,7 @@ use std::time::Instant;
 
 use halo2curves::secp256k1::{Fq as secp256k1Scalar, Secp256k1};
 use midnight_circuits::{
-    compact_std_lib::{self, Relation, ShaTableSize, ZkStdLib, ZkStdLibArch},
+    compact_std_lib::{self, Relation, ZkStdLib, ZkStdLibArch},
     field::foreign::{params::MultiEmulationParams, AssignedField},
     instructions::{
         ArithInstructions, AssertionInstructions, AssignmentInstructions, Base64Instructions,
@@ -134,7 +134,7 @@ impl Relation for AtalaJsonECDSA {
         ZkStdLibArch {
             jubjub: false,
             poseidon: false,
-            sha256: Some(ShaTableSize::Table11),
+            sha256: true,
             secp256k1: true,
             bls12_381: false,
             base64: true,