+++ circuits/examples/native_gadget.rs
@@ -92,7 +92,7 @@ impl Relation for NativeGadgetExample {
         ZkStdLibArch {
             jubjub: false,
             poseidon: false,
-            sha256: None,
+            sha256: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,