+++ circuits/examples/poseidon.rs
@@ -44,7 +44,7 @@ impl Relation for PoseidonExample {
         ZkStdLibArch {
             jubjub: false,
             poseidon: true,
-            sha256: None,
+            sha256: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,