+++ circuits/examples/rsa_signature.rs
@@ -78,7 +78,7 @@ impl Relation for RSASignatureCircuit {
         ZkStdLibArch {
             jubjub: false,
             poseidon: false,
-            sha256: None,
+            sha256: false,
             secp256k1: false,
             bls12_381: false,
             base64: false,