+++ circuits/src/compact_std_lib.rs
@@ -33,7 +33,7 @@ use group::{prime::PrimeCurveAffine, Group};
 use halo2curves::secp256k1::{self, Secp256k1};
 use midnight_curves::{G1Affine, G1Projective};
 use midnight_proofs::{
-    circuit::{Chip, Layouter, SimpleFloorPlanner, Value},
+    circuit::{Layouter, SimpleFloorPlanner, Value},
     dev::cost_model::{from_circuit_to_circuit_model, CircuitModel},
     plonk::{k_from_circuit, prepare, Circuit, ConstraintSystem, Error, ProvingKey, VerifyingKey},
     poly::{
@@ -69,10 +69,7 @@ use crate::{
     },
     hash::{
         poseidon::{PoseidonChip, PoseidonConfig, NB_POSEIDON_ADVICE_COLS, NB_POSEIDON_FIXED_COLS},
-        sha256::{
-            Sha256, Table11Chip, Table11Config, Table16Chip, Table16Config, NB_TABLE11_ADVICE_COLS,
-            NB_TABLE11_FIXED_COLS,
-        },
+        sha256::{Sha256Chip, Sha256Config, NB_SHA256_ADVICE_COLS, NB_SHA256_FIXED_COLS},
     },
     instructions::{
         hash_to_curve::HashToCurveInstructions, public_input::CommittedInstanceInstructions,
@@ -95,9 +92,6 @@ use crate::{
     vec::{vector_gadget::VectorGadget, AssignedVector, Vectorizable},
 };
 
-const SHA256_SIZE_IN_WORDS: usize = 8;
-const SHA256_SIZE_IN_BYTES: usize = 4 * SHA256_SIZE_IN_WORDS;
-
 type C = midnight_curves::JubjubExtended;
 type F = midnight_curves::Fq;
 
@@ -108,16 +102,6 @@ type Secp256k1ScalarChip = FieldChip<F, secp256k1::Fq, MEP, NG>;
 type Secp256k1Chip = ForeignEccChip<F, Secp256k1, MEP, Secp256k1ScalarChip, NG>;
 type Bls12381BaseChip = FieldChip<F, midnight_curves::Fp, MEP, NG>;
 type Bls12381Chip = ForeignEccChip<F, midnight_curves::G1Projective, MEP, NG, NG>;
-
-/// Size of the lookup table for SHA.
-#[derive(Clone, Copy, Debug, Encode, Decode)]
-pub enum ShaTableSize {
-    /// Table of size 2^11.
-    Table11,
-
-    /// Table of size 2^16.
-    Table16,
-}
 
 const ZKSTD_VERSION: u32 = 1;
 
@@ -133,8 +117,8 @@ pub struct ZkStdLibArch {
     /// Enable the Poseidon chip?
     pub poseidon: bool,
 
-    /// Enable the SHA256 chip? Using Table11 or Table16?
-    pub sha256: Option<ShaTableSize>,
+    /// Enable the SHA256 chip?
+    pub sha256: bool,
 
     /// Enable the Secp256k1 chip?
     pub secp256k1: bool,
@@ -157,7 +141,7 @@ impl Default for ZkStdLibArch {
         ZkStdLibArch {
             jubjub: true,
             poseidon: true,
-            sha256: Some(ShaTableSize::Table11),
+            sha256: true,
             secp256k1: false,
             bls12_381: false,
             base64: false,
@@ -199,8 +183,7 @@ pub struct ZkStdLibConfig {
     native_config: NativeConfig,
     core_decomposition_config: P2RDecompositionConfig,
     jubjub_config: Option<EccConfig>,
-    table11_config: Option<Table11Config>,
-    table16_config: Option<Table16Config>,
+    sha256_config: Option<Sha256Config>,
     poseidon_config: Option<PoseidonConfig<midnight_curves::Fq>>,
     secp256k1_scalar_config: Option<FieldChipConfig>,
     secp256k1_config: Option<ForeignEccConfig<Secp256k1>>,
@@ -216,8 +199,7 @@ pub struct ZkStdLib {
     native_gadget: NG,
     core_decomposition_chip: P2RDecompositionChip<F>,
     jubjub_chip: Option<EccChip<C>>,
-    sha256_table11_chip: Option<Table11Chip<F>>,
-    sha256_table16_chip: Option<Table16Chip<F>>,
+    sha256_chip: Option<Sha256Chip<F>>,
     poseidon_gadget: Option<PoseidonChip<F>>,
     htc_gadget: Option<HashToCurveGadget<F, C, AssignedNative<F>, PoseidonChip<F>, EccChip<C>>>,
     map_gadget: Option<MapGadget<F, NG, PoseidonChip<F>>>,
@@ -244,10 +226,8 @@ impl ZkStdLib {
         let native_gadget = NativeGadget::new(core_decomposition_chip.clone(), native_chip.clone());
         let jubjub_chip = (config.jubjub_config.as_ref())
             .map(|jubjub_config| EccChip::new(jubjub_config, &native_gadget));
-        let sha256_table11_chip = (config.table11_config.as_ref())
-            .map(|table11_config| Table11Chip::construct(table11_config.clone()));
-        let sha256_table16_chip = (config.table16_config.as_ref())
-            .map(|table16_config| Table16Chip::construct(table16_config.clone()));
+        let sha256_chip = (config.sha256_config.as_ref())
+            .map(|sha256_config| Sha256Chip::new(sha256_config, &native_gadget));
         let poseidon_gadget = (config.poseidon_config.as_ref())
             .map(|poseidon_config| PoseidonChip::new(poseidon_config, &native_chip));
         let htc_gadget = (jubjub_chip.as_ref())
@@ -279,8 +259,7 @@ impl ZkStdLib {
             native_gadget,
             core_decomposition_chip,
             jubjub_chip,
-            sha256_table11_chip,
-            sha256_table16_chip,
+            sha256_chip,
             poseidon_gadget,
             map_gadget,
             htc_gadget,
@@ -307,10 +286,10 @@ impl ZkStdLib {
             } else {
                 0
             },
-            match arch.sha256 {
-                Some(ShaTableSize::Table11) => NB_TABLE11_ADVICE_COLS,
-                Some(ShaTableSize::Table16) => 0, // Table16 advice cols are not shareable
-                None => 0,
+            if arch.sha256 {
+                NB_SHA256_ADVICE_COLS
+            } else {
+                0
             },
             if arch.secp256k1 {
                 max(
@@ -351,11 +330,7 @@ impl ZkStdLib {
             } else {
                 0
             },
-            match arch.sha256 {
-                Some(ShaTableSize::Table11) => NB_TABLE11_FIXED_COLS,
-                Some(ShaTableSize::Table16) => 0, // Table16 fixed cols are not shareable
-                None => 0,
-            },
+            if arch.sha256 { NB_SHA256_FIXED_COLS } else { 0 },
         ]
         .into_iter()
         .max()
@@ -393,17 +368,14 @@ impl ZkStdLib {
             false => None,
         };
 
-        let table11_config = match arch.sha256 {
-            Some(ShaTableSize::Table11) => Some(Table11Chip::configure(
+        let sha256_config = match arch.sha256 {
+            true => Some(Sha256Chip::configure(
                 meta,
-                &advice_columns[..NB_TABLE11_ADVICE_COLS].try_into().unwrap(),
-                fixed_columns[..NB_TABLE11_FIXED_COLS].try_into().unwrap(),
+                &(
+                    advice_columns[..NB_SHA256_ADVICE_COLS].try_into().unwrap(),
+                    fixed_columns[..NB_SHA256_FIXED_COLS].try_into().unwrap(),
+                ),
             )),
-            _ => None,
-        };
-
-        let table16_config = match arch.sha256 {
-            Some(ShaTableSize::Table16) => Some(Table16Chip::configure(meta)),
             _ => None,
         };
 
@@ -473,8 +445,7 @@ impl ZkStdLib {
             native_config,
             core_decomposition_config,
             jubjub_config,
-            table11_config,
-            table16_config,
+            sha256_config,
             poseidon_config,
             secp256k1_scalar_config,
             secp256k1_config,
@@ -550,7 +521,7 @@ impl ZkStdLib {
     /// Assert that a given assigned bit is true.
     ///
     /// ```
-    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 12, {
+    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 13, {
     /// let input: AssignedBit<F> = chip.assign_fixed(layouter, true)?;
     /// chip.assert_true(layouter, &input)?;
     /// # });
@@ -577,7 +548,7 @@ impl ZkStdLib {
     /// Returns `1` iff `x < y`.
     ///
     /// ```
-    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 12, {
+    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 13, {
     /// let x: AssignedNative<F> = chip.assign_fixed(layouter, F::from(127))?;
     /// let y: AssignedNative<F> = chip.assign_fixed(layouter, F::from(212))?;
     /// let condition = chip.lower_than(layouter, &x, &y, 8)?;
@@ -592,7 +563,7 @@ impl ZkStdLib {
     /// condition is violated, the circuit becomes unsatisfiable.
     ///
     /// ```should_panic
-    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 12, {
+    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 13, {
     /// let x: AssignedNative<F> = chip.assign_fixed(layouter, F::from(127))?;
     /// let y: AssignedNative<F> = chip.assign_fixed(layouter, F::from(212))?;
     /// let _condition = chip.lower_than(layouter, &x, &y, 7)?;
@@ -621,7 +592,7 @@ impl ZkStdLib {
     /// Poseidon hash from a slice of native valure into a native value.
     ///
     /// ```
-    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 12, {
+    /// # midnight_circuits::run_test_std_lib!(chip, layouter, 13, {
     /// let x: AssignedNative<F> = chip.assign_fixed(layouter, F::from(127))?;
     /// let y: AssignedNative<F> = chip.assign_fixed(layouter, F::from(212))?;
     ///
@@ -675,22 +646,12 @@ impl ZkStdLib {
         &self,
         layouter: &mut impl Layouter<F>,
         input: &[AssignedByte<F>], // F -> decompose_bytes -> hash
-    ) -> Result<[AssignedByte<F>; SHA256_SIZE_IN_BYTES], Error> {
+    ) -> Result<[AssignedByte<F>; 32], Error> {
         *self.used_sha.borrow_mut() = true;
-
-        // Note that both table11 and table16 SHA chips cannot be enabled at the same
-        // time, but one must be enabled.
-        if let Some(sha256_chip) = self.sha256_table11_chip.as_ref() {
-            let hasher = Sha256::new(sha256_chip.clone(), self.native_gadget.clone())?;
-            return hasher.hash(layouter, input);
-        }
-
-        if let Some(sha256_chip) = self.sha256_table16_chip.as_ref() {
-            let hasher = Sha256::new(sha256_chip.clone(), self.native_gadget.clone())?;
-            return hasher.hash(layouter, input);
-        }
-
-        panic!("ZkStdArch must enable sha256 (as Table11 or Table16)")
+        self.sha256_chip
+            .as_ref()
+            .expect("ZkStdArch must enable sha256")
+            .hash(layouter, input)
     }
 }
 
@@ -1301,16 +1262,15 @@ impl<Rel: Relation> MidnightPK<Rel> {
 /// # Example
 ///
 /// ```
-/// # use midnight_curves::G1Affine;
-/// # use midnight_proofs::{
-/// #     circuit::{Layouter, Value},
-/// #     plonk::Error,
-/// # };
 /// # use midnight_circuits::{
-/// #     compact_std_lib::{self, MidnightCircuit, Relation, ShaTableSize, ZkStdLib, ZkStdLibArch},
+/// #     compact_std_lib::{self, Relation, ZkStdLib, ZkStdLibArch},
 /// #     instructions::{AssignmentInstructions, PublicInputInstructions},
 /// #     testing_utils::plonk_api::filecoin_srs,
 /// #     types::{AssignedByte, Instantiable},
+/// # };
+/// # use midnight_proofs::{
+/// #     circuit::{Layouter, Value},
+/// #     plonk::Error,
 /// # };
 /// # use rand::{rngs::OsRng, Rng, SeedableRng};
 /// # use rand_chacha::ChaCha8Rng;
@@ -1318,7 +1278,7 @@ impl<Rel: Relation> MidnightPK<Rel> {
 /// #
 /// type F = midnight_curves::Fq;
 ///
-/// #[derive(Clone)]
+/// #[derive(Clone, Default)]
 /// struct ShaPreImageCircuit;
 ///
 /// impl Relation for ShaPreImageCircuit {
@@ -1353,11 +1313,11 @@ impl<Rel: Relation> MidnightPK<Rel> {
 ///     }
 ///
 ///     fn write_relation<W: std::io::Write>(&self, _writer: &mut W) -> std::io::Result<()> {
-///        Ok(())
+///         Ok(())
 ///     }
 ///
 ///     fn read_relation<R: std::io::Read>(_reader: &mut R) -> std::io::Result<Self> {
-///        Ok(ShaPreImageCircuit)
+///         Ok(ShaPreImageCircuit)
 ///     }
 /// }
 ///
@@ -1377,6 +1337,7 @@ impl<Rel: Relation> MidnightPK<Rel> {
 ///     &srs, &pk, &relation, &instance, witness, OsRng,
 /// )
 /// .expect("Proof generation should not fail");
+///
 /// assert!(
 ///     compact_std_lib::verify::<ShaPreImageCircuit, blake2b_simd::State>(
 ///         &srs.verifier_params(),
@@ -1422,7 +1383,7 @@ pub trait Relation: Clone {
     /// The blanket implementation enables:
     ///  - jubjub
     ///  - poseidon
-    ///  - sha256 (with 2^11 table)
+    ///  - sha256
     fn used_chips(&self) -> ZkStdLibArch {
         ZkStdLibArch::default()
     }
@@ -1463,11 +1424,7 @@ impl<R: Relation> Circuit<F> for MidnightCircuit<'_, R> {
         config: Self::Config,
         mut layouter: impl Layouter<F>,
     ) -> Result<(), Error> {
-        let max_bit_len = match config.architecture.sha256 {
-            None => 8,
-            Some(ShaTableSize::Table11) => 10,
-            Some(ShaTableSize::Table16) => 15,
-        };
+        let max_bit_len = if config.architecture.sha256 { 11 } else { 8 };
         let zk_std_lib = ZkStdLib::new(&config, max_bit_len);
 
         self.relation.circuit(
@@ -1491,13 +1448,9 @@ impl<R: Relation> Circuit<F> for MidnightCircuit<'_, R> {
             b64_chip.load(&mut layouter)?;
         }
 
-        if *zk_std_lib.used_sha.borrow() {
-            if let Some(sha256_chip) = zk_std_lib.sha256_table11_chip {
-                Table11Chip::load(sha256_chip.config().clone(), &mut layouter)?
-            }
-
-            if let Some(sha256_chip) = zk_std_lib.sha256_table16_chip {
-                Table16Chip::load(sha256_chip.config().clone(), &mut layouter)?
+        if let Some(sha256_chip) = zk_std_lib.sha256_chip {
+            if *zk_std_lib.used_sha.borrow() {
+                sha256_chip.load(&mut layouter)?;
             }
         }