+++ circuits/CHANGELOG.md
@@ -14,6 +14,7 @@ verification keys break backwards compatibility.
 ### Added
 * Add CommittedInstanceInstructions [#63](https://github.com/midnightntwrk/midnight-zk/pull/63)
 * Add goldenfiles with examples cost-model [#89](https://github.com/midnightntwrk/midnight-zk/pull/89)
+* New SHA256 chip [#39](https://github.com/midnightntwrk/midnight-zk/pull/39)
 
 ### Changed
 * Refactored crate to prepare for edition 2024.