+++ proofs/CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+* Implement `From<u64>` for Expression[#39](https://github.com/midnightntwrk/midnight-zk/pull/39)
 ### Changed
 * API for defining custom constraints was unified [#53](https://github.com/midnightntwrk/midnight-zk/pull/53)
 ### Removed